requirement.txt
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
google-generativeai
pypdf
chromadb
python-dotenv
cryptography
sentence-transformers